# docs/swagger.yaml
definitions:
  models.BalanceTransaction:
    properties:
      balance:
        type: number
      quantity:
        type: integer
    type: object
  models.CloseShipping:
    properties:
      id_shipping:
        type: integer
    required:
    - id_shipping
    type: object
  models.CreateShipping:
    properties:
      amount:
        type: number
      expiration_date:
        type: string
      id_user:
        type: integer
      quantity:
        type: integer
      shipping_number:
        type: string
      username:
        type: string
      weight:
        type: number
    required:
    - amount
    - expiration_date
    - quantity
    - shipping_number
    - username
    - weight
    type: object
  models.CreateTransaction:
    properties:
      amount:
        type: number
      id_shipping:
        type: integer
      id_transaction_type:
        type: integer
      id_user:
        type: integer
    required:
    - amount
    - id_shipping
    - id_transaction_type
    - id_user
    type: object
  models.ErrorMessage:
    properties:
      error:
        type: string
    type: object
  models.LoginUser:
    properties:
      password:
        type: string
      token:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.RegisterUser:
    properties:
      cc:
        type: string
      direction:
        type: string
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        minLength: 8
        type: string
      passwordc:
        type: string
      phonenumber:
        type: string
      username:
        type: string
    required:
    - cc
    - email
    - firstname
    - password
    - passwordc
    - phonenumber
    - username
    type: object
  models.Shipping:
    properties:
      amount:
        type: number
      created_date:
        type: string
      email:
        type: string
      expiration_date:
        type: string
      id_shipping:
        type: integer
      id_user:
        type: integer
      last_update:
        type: string
      quantity:
        type: integer
      shipping_number:
        type: string
      status:
        type: integer
      username:
        type: string
      weight:
        type: number
    type: object
  models.SuccessMessage:
    properties:
      message:
        type: string
    type: object
  models.UpdateShipping:
    properties:
      amount:
        type: number
      expiration_date:
        type: string
      id_shipping:
        type: integer
      quantity:
        type: integer
      shipping_number:
        type: string
      status:
        type: integer
      username:
        type: string
      weight:
        type: number
    required:
    - id_shipping
    type: object
  models.User:
    properties:
      cc:
        type: string
      created_date:
        type: string
      direction:
        type: string
      email:
        type: string
      firstname:
        type: string
      id_role:
        type: integer
      id_user:
        type: integer
      lastname:
        type: string
      phone_number:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  title: Tag Service API
paths:
  /shipping:
    get:
      description: Realiza la busqueda de pedidos del usuario por defecto o email,
        id de paquete o número de paquete
      parameters:
      - description: Correo electrónico del usuario a buscar
        in: query
        name: email
        type: string
      - description: Código de paquete a buscar
        in: query
        name: shipping_number
        type: integer
      - description: Id de paquete a buscar
        in: query
        name: id_shipping
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de paquetes encontrados
          schema:
            items:
              $ref: '#/definitions/models.Shipping'
            type: array
        "400":
          description: Error en los datos proporcionados
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      security:
      - JWT: []
      summary: Busqueda de paquetes
      tags:
      - Shipping
    post:
      consumes:
      - application/json
      description: Realiza el guardado del paquete que se quiere crear en la BD y
        una transacción se abre
      parameters:
      - description: crea paquete y transacción
        in: body
        name: shipping
        required: true
        schema:
          $ref: '#/definitions/models.CreateShipping'
      produces:
      - application/json
      responses:
        "200":
          description: mensaje de éxito
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Error en los datos proporcionados
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      security:
      - JWT: []
      summary: Creación de paquete
      tags:
      - Shipping
    put:
      consumes:
      - application/json
      parameters:
      - description: actualiza los datos del paquete
        in: body
        name: shipping
        required: true
        schema:
          $ref: '#/definitions/models.UpdateShipping'
      produces:
      - application/json
      responses:
        "200":
          description: mensaje de éxito
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Error en el proceso
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      security:
      - JWT: []
      summary: Actualización de datos de paquete
      tags:
      - Shipping
  /shipping/close:
    post:
      consumes:
      - application/json
      parameters:
      - description: toma el id del paquete para cerrarlo y crear transacción de cierre
        in: body
        name: shipping
        required: true
        schema:
          $ref: '#/definitions/models.CloseShipping'
      produces:
      - application/json
      responses:
        "200":
          description: mensaje de éxito
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Error en el proceso
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      security:
      - JWT: []
      summary: Cerrar paquete y creación de transacción
      tags:
      - Shipping
  /transaction:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ESTADO DE SALDO Y PAQUETES QUE TIENE
          schema:
            $ref: '#/definitions/models.BalanceTransaction'
        "400":
          description: Error en los datos proporcionados
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      security:
      - JWT: []
      summary: Estado de cuenta de usuario
      tags:
      - Transaction
    post:
      consumes:
      - application/json
      description: Realiza el guardado de la transacción que se quiere crear en la
        BD
      parameters:
      - description: crea transacción
        in: body
        name: shipping
        required: true
        schema:
          $ref: '#/definitions/models.CreateTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: mensaje de éxito
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Error en los datos proporcionados
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      security:
      - JWT: []
      summary: Creación de transacción
      tags:
      - Transaction
  /user:
    post:
      consumes:
      - application/json
      description: Realiza el guardado del usuario en la BD
      parameters:
      - description: crea usuario
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/models.RegisterUser'
      produces:
      - application/json
      responses:
        "200":
          description: mensaje de éxito
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Error en los datos proporcionados
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: Registro de usuario
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: Realiza el guardado del usuario en la BD
      parameters:
      - description: Email y contraseña
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/models.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: mensaje de éxito
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Error en los datos proporcionados
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: Inicio de sesión de usuario
      tags:
      - User
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"